+++ composeApp/src/commonMain/kotlin/feature/news/data/api/NewsDetailApiServiceImpl.kt
@@ -12,7 +12,8 @@ class NewsDetailApiServiceImpl(
     private val httpClient: HttpClient,
     private val appConfig: AppConfig
 ) : NewsDetailApiService {
-    override suspend fun getNewsDetail(id: Int): BaseResponse<ArticleDetailResponse> {
+
+    override suspend fun getNewsDetail(id: Long): BaseResponse<ArticleDetailResponse>{
         val xid = id % 8
         val url = "${appConfig.baseApiUrl}details"
         return httpClient.get(url) {
