+++ composeApp/src/commonMain/kotlin/feature/news/domain/repository/NewsDetailRepository.kt
@@ -5,6 +5,6 @@ import feature.news.domain.model.NewsDetail
 import kotlinx.coroutines.flow.Flow
 
 interface NewsDetailRepository {
-    fun getNewsDetail(id: Int): Flow<NewsDetail?>
-    suspend fun refreshNewsDetail(id: Int): Result<Unit>
+    fun getNewsDetail(id: Long): Flow<NewsDetail?>
+    suspend fun refreshNewsDetail(id: Long): Result<Unit>
 }
