+++ composeApp/src/commonMain/kotlin/feature/news/domain/usecase/newsdetail/RefreshNewsDetailUseCase.kt
@@ -4,7 +4,7 @@ import core.domain.model.Result
 import feature.news.domain.repository.NewsDetailRepository
 
 class RefreshNewsDetailUseCase(private val repository: NewsDetailRepository) {
-    suspend operator fun invoke(id: Int): Result<Unit> {
+    suspend operator fun invoke(id: Long): Result<Unit> {
         return repository.refreshNewsDetail(id)
     }
 }
